backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
Pillow==10.4.0
python-decouple==3.8
whitenoise==6.8.2
